:tractor: Split out parsing into its own separate files for better unit test visibility
